Update Troubleshooting remove Static IPv6
A static IPv6 Address for the HA network interface is not supported with Thread networks. ICMPv6 Router Advertisements, with routes to the Thread Devices, will not be stored in the Host routing table when a static IPv6 Address is configured. The suggestion to use a static IPv6 Address is therefore removed.

(My first commit ever, I hope it's okay)